Update shop.html
This code wraps the product image in a link (<a>) tag with the data-lightbox attribute. This attribute tells the lightbox library to open the linked image in a modal window when clicked. The href attribute of the link points to the full-size image to be displayed in the lightbox.

This code includes the fslightbox library from Cloudflare.